requirements.txt
--find-links https://download.pytorch.org/whl/torch
torch==2.5.1+cu124
torchvision
torchaudio
cupy-cuda12x
structure-tensor
monai
matplotlib==3.9.0
torchio
lightning
timm
kornia
opencv-python
wandb
hydra-core
invoke
scikit-learn
scikit-image
matplotlib
napari
pipreqs
typer>=0.15.1
fastapi>=0.115.6
uvicorn>=0.34.0
coverage>=6.0.2
pytest>=6.2.4
ruff
python-multipart
httpx
locust
netron
bentoml
evidently
pandas
h5py
basicsr
einops